Move util into sbt.util.internal package
