+++ internal/util-control/src/main/scala/sbt/util/internal/ExitHook.scala
@@ -1,7 +1,7 @@
 /* sbt -- Simple Build Tool
  * Copyright 2009, 2010  Mark Harrah
  */
-package sbt
+package sbt.util.internal
 
 /** Defines a function to call as sbt exits.*/
 trait ExitHook {
