+++ internal/util-tracking/src/main/scala/sbt/util/internal/ChangeReport.scala
@@ -1,7 +1,7 @@
 /* sbt -- Simple Build Tool
  * Copyright 2009, 2010 Mark Harrah
  */
-package sbt
+package sbt.util.internal
 
 object ChangeReport {
   def modified[T](files: Set[T]): ChangeReport[T] =
